c2: add a predefined target for whether a window is on top of others
The new 'occluding_a_window' predef target allows the user to express
the aforementioned condition, which is very useful for making
normally-transparent windows opaque when they are on top of another
window.
This is especially useful for situations where a floating terminal
wants to be transparent, but not when the transparency would make the
text unreadable.